requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
requests==2.31.0
python-decouple==3.8
psycopg2-binary==2.9.9
django-extensions==3.2.3
whitenoise==6.6.0
pytz==2023.3
